Separate feature selection into distinct submodules
