+++ divik/_feature_selection/__init__.py
@@ -1,0 +1,2 @@
+from ._gmm_selector import GMMSelector
+from ._specialized import HighAbundanceAndVarianceSelector
